pyproject.toml
[project]
name = "drf-timezone-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "django-filter>=25.2",
    "django-rest-framework>=0.1.0",
    "djangorestframework>=3.16.1",
    "psycopg>=3.2.10",
    "tzlocal>=5.3.1",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
]
